// packages/backend.ai-ui/src/__generated__/BAIUserSelectAdminPaginatedQuery.graphql.ts
export type UserV2Filter = {
  AND?: ReadonlyArray<UserV2Filter> | null | undefined;
  NOT?: ReadonlyArray<UserV2Filter> | null | undefined;
  OR?: ReadonlyArray<UserV2Filter> | null | undefined;
  containerGids?: IntArrayFilter | null | undefined;
  containerMainGid?: IntFilter | null | undefined;
  containerUid?: IntFilter | null | undefined;
  createdAt?: DateTimeFilter | null | undefined;
  description?: StringFilter | null | undefined;
  domain?: UserDomainNestedFilter | null | undefined;
  domainId?: UUIDFilter | null | undefined;
  domainName?: StringFilter | null | undefined;
  email?: StringFilter | null | undefined;
  fullName?: StringFilter | null | undefined;
  integrationName?: StringFilter | null | undefined;
  keypairs?: UserKeypairNestedFilter | null | undefined;
  modifiedAt?: DateTimeFilter | null | undefined;
  needPasswordChange?: boolean | null | undefined;
  project?: UserProjectNestedFilter | null | undefined;
  resourcePolicy?: StringFilter | null | undefined;
  role?: UserRoleV2EnumFilter | null | undefined;
  status?: UserStatusV2EnumFilter | null | undefined;
  statusInfo?: StringFilter | null | undefined;
  sudoSessionEnabled?: boolean | null | undefined;
  totpActivated?: boolean | null | undefined;
  totpActivatedAt?: NullableDateTimeFilter | null | undefined;
  username?: StringFilter | null | undefined;
  uuid?: UUIDFilter | null | undefined;
};
// ...
export type UserV2OrderBy = {
  direction?: OrderDirection;
  field?: UserV2OrderField;
};
export type BAIUserSelectAdminPaginatedQuery$variables = {
  filter?: UserV2Filter | null | undefined;
  limit: number;
  offset: number;
  orderBy?: ReadonlyArray<UserV2OrderBy> | null | undefined;
};
export type BAIUserSelectAdminPaginatedQuery$data = {
  readonly adminUsersV2: {
    readonly count: number;
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly basicInfo: {
          readonly email: string;
          readonly fullName: string | null | undefined;
        };
        readonly id: string;
      };
    }>;
  } | null | undefined;
};
export type BAIUserSelectAdminPaginatedQuery = {
  response: BAIUserSelectAdminPaginatedQuery$data;
  variables: BAIUserSelectAdminPaginatedQuery$variables;
};